Feat[cropper]: DO NOT USE. Begin implementing support for running with BitmapRegionDecoder
TODO:
- actually add cropping to the BitmapRegionDecoder code
- implement out-of-bounds region decoding
